pkg/html: split off URL helpers
The existing pkg/html package transitively takes too many dependencies,
which complicates the reuse of some of its methods.

Split off a pkg/html/urlutil package.
